feat: move curriculum list components to Home
