+++ frontend/src/components/Home/AISuggestion.tsx
