+++ frontend/src/components/Home/AddCurriculumCard.tsx
