+++ frontend/src/components/Home/Card.tsx
